requirements.txt
Django==1.11.3
django-ckeditor==5.3.0
django-js-asset==0.1.1
gunicorn==19.7.1
psycopg2==2.7.1
pytz==2017.2
requests==2.14.2
whitenoise==3.3.0
